refactor: split endpoint reader and writer
