+++ logs/log1756126139.md
@@ -1,0 +1,5 @@
+## Log 1756126139
+- Split connection reader and writer loops into dedicated `ConnectionReader` and `ConnectionWriter` types under `Endpoints/Runner` to improve separation of concerns and testability.
+- Updated `ConnectionRunner` to orchestrate the new types and moved it into the new `Runner` directory.
+- Verified .NET SDK installation and rebuilt solution.
+- Ran core test suites; `Proto.Actor.Tests` and `Proto.Remote.Tests` passed, while `Proto.Cluster.Tests` had one failing test (`GossipCoreTests.Large_cluster_should_get_topology_consensus`).
